# Restore more efficient version of NewPossibleNonCounterInfo annotation