Follow up to https://github.com/prometheus/prometheus/pull/13012
- Put back NewPossibleNonCounterInfo and run it only once per series instead of at every step
- ~Run NewInvalidQuantileWarning only once per query instead of at every step~

Benchmark for overhead for rate queries (before is run on [a branch on main that reverts the commits that added and modified annotations](https://github.com/zenador/prometheus/tree/temp-benchmark-remove-annotations), and after is run on this branch):
```
                                                  │ BenchmarkRangeQuery_before.txt │    BenchmarkRangeQuery_after.txt    │
                                                  │             sec/op             │    sec/op     vs base               │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                     9.345m ± 18%   10.076m ± 3%  +7.83% (p=0.029 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │    BenchmarkRangeQuery_after.txt    │
                                                  │              B/op              │     B/op      vs base               │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                     367.6Ki ± 0%   402.6Ki ± 1%  +9.54% (p=0.000 n=10)

                                                  │ BenchmarkRangeQuery_before.txt │    BenchmarkRangeQuery_after.txt    │
                                                  │           allocs/op            │  allocs/op   vs base                │
RangeQuery/expr=rate(a_hundred[1m]),steps=1000-16                      5.256k ± 0%   5.859k ± 0%  +11.47% (p=0.000 n=10)
```