@zenador 2023-12-31
Good point, I didn't consider the case where quantile value could change if it's based on another metric instead of a static expression. In that case, we do have to run it every step, and the deduplication is already done, so I reverted those commits and this PR is now only addressing `NewPossibleNonCounterInfo`.